backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Compress==1.13
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
reportlab==4.0.8
qrcode==7.4.2
Pillow==10.1.0 
uuid
datetime
# Nowe pakiety dla Import/Export API v2.0
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9 
# Nowe pakiety dla Import/Export API v2.0
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
